refactoring to move most functionality to the gtex utility class.
gtex_builder now only contains functionality to load vars from a file.
